requirements.txt
facebook-business>=19.0.0
python-dotenv>=1.0.0
openai>=1.30.0
Pillow>=10.0.0
APScheduler>=3.10.0
streamlit>=1.30.0
